feat(add-service-adapapter): update matching response

+++ common/src/main/java/org/hikit/common/dto/MatchingResponse.kt
@@ -5,5 +5,5 @@ data class MatchingResponse (
 )
 
 data class TrailToScore(val trailData: Trail,
-                        val accuracy: Int)
+                        val accuracy: Double)
